requirements.txt
cachetools==5.3.0
charset-normalizer==3.0.1
click==8.1.3
dnspython==2.3.0
Flask==2.2.2
Flask-Cors==3.0.10
google-auth==2.16.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kubernetes==26.1.0
loguru==0.6.0
MarkupSafe==2.1.2
numpy==1.24.2
oauthlib==3.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==4.3.3
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
threadpoolctl==3.1.0
urllib3==1.26.14
websocket-client==1.5.1
Werkzeug==2.2.2
zipp==3.12.0
pandas~=1.5.3
pytz~=2022.7
apscheduler~=3.9.0
grpcio==1.70.0
grpcio-tools==1.70.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
colorama==0.4.6
distro==1.9.0
frozenlist==1.6.0
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.58
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
langgraph==0.4.1
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.66
langsmith==0.3.42
multidict==6.4.3
openai==1.77.0
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
propcache==0.3.1
pydantic==2.11.4
pydantic_core==2.33.2
regex==2024.11.6
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.40
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
contourpy==1.3.0
cycler==0.12.1
fonttools==4.58.0
kiwisolver==1.4.7
matplotlib==3.9.4
networkx==3.2.1
pillow==9.5.0
pyparsing==3.0.9
requests==2.28.2